Looking at other languages that allow broad unicode use in variable assignment (such as Swift), I came up with a blacklist of certain carhacters we should avoid.
This doesn't go as far as some other languages: for example, I'm not
including mathematical symbols, or punctuation in general. The "ASCII
Punctuation & Symbols" table in unicode, which is a superset of most of
these reserved characters, contains others that could potentially be
useful, such as `$` for representing jQuery.

whitespace
  0000 null
  0009 horizontal tab
  000A line feed
  000B vertical tab
  000C form feed
  000D carriage return
  0020 space
  FEFF BOM

AML reserved characters
  005C \
  005B [
  005D ]
  007B {
  007D }
  003A :

  (allowed in slug matching, but interpreted outside of the slug)
  002E .
  002B + (in use in the freeform markup branch; possibly only when leading, or trailing?)

Possibly reserve for future use?
  005E ^
  0060 `
  007C |
  007E ~